refactor: transaction voted for handling (#1225)
