+++ app/Http/Livewire/WalletTransactionTable.php
@@ -161,6 +161,7 @@ final class WalletTransactionTable extends TabbedTableComponent
     {
         return Transaction::query()
             ->withTypeFilter($this->filter)
+            ->with('votedFor')
             ->where(function ($query) {
                 $query->where(fn ($query) => $query->when($this->filter['outgoing'], fn ($query) => $query->where('sender_public_key', $this->publicKey)))
                     ->orWhere(fn ($query) => $query->when($this->filter['incoming'], fn ($query) => $query->where('to', $this->address)))
