+++ tests/Unit/ViewModels/TransactionViewModelTest.php
@@ -417,14 +417,14 @@ MethodID: 0x6dd7d8ea');
         ]);
     });
 
-    it('should fail to get formatted multi payment receipts if not a multi payment', function () {
+    it('should fail to get formatted multi payment recipients if invalid a multi payment', function () {
         $transaction = new TransactionViewModel(Transaction::factory()
             ->withPayload('123456')
             ->create());
 
         expect(function () use ($transaction) {
             $transaction->multiPaymentRecipients();
-        })->toThrow(Exception::class);
+        })->toThrow(Exception::class, 'This transaction is not a multi-payment.');
     });
 });
